fix: remove ios scroll reset workaround (#127)
Pulling down to refresh does not seem to work on the Example app using Expo, when the tab pane is scrolled to the top.

This only seems to happen in that case though, in a bare RN iOS app everything seems fine.

This workaround creates other issue though, and isn't correct. We'll need to find a better solution.